+++ src/pykeen/metrics/utils.py
@@ -5,6 +5,7 @@
 from dataclasses import dataclass
 from typing import ClassVar, Collection, Iterable, Optional
 
+import numpy as np
 from docdata import get_docdata
 
 from ..utils import camel_to_snake
@@ -12,6 +13,9 @@ from ..utils import camel_to_snake
 __all__ = [
     "Metric",
     "ValueRange",
+    "stable_product",
+    "weighted_mean_expectation",
+    "weighted_mean_variance",
 ]
 
 
@@ -138,3 +142,27 @@ class Metric:
 
     def __repr__(self) -> str:  # noqa:D105
         return f"{self.__class__.__name__}({', '.join(self._extra_repr())})"
+
+
+def weighted_mean_expectation(individual: np.ndarray, weights: Optional[np.ndarray]) -> float:
+    """Calculate the expectation of a weighted sum of variables with given individual expected value."""
+    return np.average(individual, weights=weights).item()
+
+
+def weighted_mean_variance(individual: np.ndarray, weights: Optional[np.ndarray]) -> float:
+    """Calculate the variance of a weighted sum of variables with given individual variances."""
+    n = individual.size
+    if weights is None:
+        return individual.mean() / n
+    weights = weights / weights.sum()
+    return (individual * weights**2).sum().item()
+
+
+def stable_product(a: np.ndarray, axis: Optional[int] = None, is_log: bool = False) -> np.ndarray:
+    """Compute product using the log-trick for increased numerical stability."""
+    if is_log:
+        sign = 1
+    else:
+        sign = np.prod(np.copysign(np.ones_like(a), a))
+        a = np.log(np.abs(a))
+    return sign * np.exp(np.sum(a, axis=axis))
